\name{HelpWithZnewgen2-package}
\alias{HelpWithZnewgen2-package}
\alias{HelpWithZnewgen2}
\docType{package}
\title{
\packageTitle{HelpWithZnewgen2}
}
\description{
\packageDescription{HelpWithZnewgen2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HelpWithZnewgen2}
\packageIndices{HelpWithZnewgen2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HelpWithZnewgen2}

Maintainer: \packageMaintainer{HelpWithZnewgen2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
